+++ src/dcos_e2e/backends/_vagrant/__init__.py
@@ -10,10 +10,9 @@ from pathlib import Path
 from tempfile import gettempdir
 from typing import Any, Dict, Iterable, Optional, Set, Tuple, Type
 
+from dcos_e2e.base_classes import ClusterBackend, ClusterManager
 from dcos_e2e.cluster import Cluster
 from dcos_e2e.node import Node, Output
-
-from .._base_classes import ClusterBackend, ClusterManager
 
 
 class Vagrant(ClusterBackend):