+++ src/dcos_e2e/cluster.py
@@ -12,13 +12,13 @@ import retrying
 import timeout_decorator
 from retry import retry
 
+from ._existing_cluster import ExistingCluster as _ExistingCluster
 from ._vendor.dcos_test_utils.dcos_api import DcosApiSession, DcosUser
 from ._vendor.dcos_test_utils.enterprise import EnterpriseApiSession
 from ._vendor.dcos_test_utils.helpers import CI_CREDENTIALS
-# Ignore a spurious error - this import is used in a type hint.
-from .backends import ClusterManager  # noqa: F401
-from .backends import ClusterBackend
 from .backends._existing_cluster import ExistingCluster as _ExistingCluster
+from .base_classes import ClusterManager  # noqa: F401
+from .base_classes import ClusterBackend
 from .exceptions import DCOSTimeoutError
 from .node import Node, Output, Transport